+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Bug fix: Automated: Eloquent
 - DowntimeUI: visual improvement and informative text
 - Automated: Soothing Speech
+- Automated: Armorer
 
 # 0.1.8
 - Added scrollbar to DowntimeUI